Backend/requirements.txt
Django==4.1.7
django-picklefield==3.1
djangorestframework==3.14.0
func_timeout==4.3.5
joblib==1.2.0
numpy==1.24.1
pandas==1.5.3
py2neo==2021.2.4
pymysql==1.0.2
scikit_learn==1.2.1
torch==1.13.1+cu116
websockets==10.4
django-redis==5.2.0
django-cors-headers==3.13.0
django-crontab==0.7.1
ipip-ipdb==1.6.1
redis==4.4.0rc4
dpkt==1.9.8
tzdata
cryptography